+++ src/main/java/dto/LottoPurchaseDto.java
@@ -2,5 +2,5 @@ package dto;
 
 import domain.Lottos;
 
-public record LottoPurchaseDto(int purchaseAmount, int manualCount, Lottos manualLottos, int autoCount) {
+public record LottoPurchaseDto(int totalAmount, int manualLottoCount, Lottos manualLottos, int autoLottoCount) {
 }
