--- src/main/java/dto/MatchResultDto.java
@@ -1,3 +1,0 @@
-package dto;
-
-public record MatchResultDto(int matchCount, boolean bonusMatch, int prize, int count) {}
