--- src/main/java/dto/ProfitDto.java
@@ -1,3 +1,0 @@
-package dto;
-
-public record ProfitDto(double rate, boolean isLoss) {}
