--- src/main/java/dto/WinningResultDto.java
@@ -1,5 +1,0 @@
-package dto;
-
-import java.util.List;
-
-public record WinningResultDto(List<MatchResultDto> matches) {}
