+++ src/test/java/domain/RankTest.java
@@ -14,11 +14,11 @@ class RankTest {
     @DisplayName("일치 개수에 따라 보너스 없이 적절한 Rank를 반환한다")
     void matchCountOf_returnsCorrectRank() {
         SoftAssertions softly = new SoftAssertions();
-        softly.assertThat(Rank.value(3, false)).isEqualTo(Rank.FIFTH);
-        softly.assertThat(Rank.value(4, false)).isEqualTo(Rank.FOURTH);
-        softly.assertThat(Rank.value(5, false)).isEqualTo(Rank.THIRD);
-        softly.assertThat(Rank.value(5, true)).isEqualTo(Rank.SECOND);
-        softly.assertThat(Rank.value(6, false)).isEqualTo(Rank.FIRST);
+        softly.assertThat(Rank.valueOf(3, false)).isEqualTo(Rank.FIFTH);
+        softly.assertThat(Rank.valueOf(4, false)).isEqualTo(Rank.FOURTH);
+        softly.assertThat(Rank.valueOf(5, false)).isEqualTo(Rank.THIRD);
+        softly.assertThat(Rank.valueOf(5, true)).isEqualTo(Rank.SECOND);
+        softly.assertThat(Rank.valueOf(6, false)).isEqualTo(Rank.FIRST);
         softly.assertAll();
     }
 
@@ -26,7 +26,7 @@ class RankTest {
     @ValueSource(ints = {0, 1, 2, 7, 10})
     @DisplayName("유효하지 않은 개수는 MISS를 반환한다")
     void matchCountOf_returnsNoneForInvalidCount(int count) {
-        assertThat(Rank.value(count, false)).isEqualTo(Rank.MISS);
+        assertThat(Rank.valueOf(count, false)).isEqualTo(Rank.MISS);
     }
 
     @Test
